Update cnv_visualising.R
First, add 'folder' argument to support users set name of a new folder to save results. Second, add code to check if the columns name in input file fit the minimum requirement. Third, revise the first condition which used to plot CNVs on all chromosome simultaneously in more clear way,  and support to adjust the size of final plots.