Comparison of CNV and CNVR between different versions was deprecated
Give up the function used for comparison between different version of reference genome assembly. Because we found that after conversion of coordinates of intervals, the total length of new intervals are differ to the old files. Because of the position of SNPs are changed among different map files, especially the order of the same SNPs on each chromosome are changed a lot! This results suggest not to use the converted intervals (CNV or CNVR) to make further comparison between various results.