Update README.md
Replace the path of 'Work Flow' figure from 'raw/' to 'blob'. Because of the graph with path in 'raw/' are only showing while browse in computer, not working in Github page and phone. The 'blob/' seems like could link the file from Github repo to 'readme.md', now it is working on computer, will check whether it will work in Github page and on phone.